# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StaxVibe AI Graphics marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  fileUrl: string | null;
  featured: boolean;
  /** @nullable */
  badge: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  featured?: boolean;
  /** @nullable */
  badge?: string | null;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  featured?: boolean;
  /** @nullable */
  badge?: string | null;
}

export interface Category {
  name: string;
  count: number;
  description: string;
}

export interface CheckoutInput {
  productId: number;
  /** @nullable */
  customerEmail?: string | null;
}

export interface CheckoutSession {
  sessionId: string;
  url: string;
}

export interface Order {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  customerEmail: string | null;
  amountPaid: number;
  status: string;
  /** @nullable */
  stripeSessionId: string | null;
  /** @nullable */
  downloadToken: string | null;
  createdAt: string;
}

export interface DownloadLink {
  url: string;
  expiresAt: string;
  productName: string;
}

export interface TopProduct {
  id: number;
  name: string;
  revenue: number;
  orders: number;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  avgOrderValue: number;
  topProducts: TopProduct[];
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export type ListProductsParams = {
  category?: string;
  featured?: string;
};