# lib/api-zod/src/generated/types/downloadLink.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StaxVibe AI Graphics marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadLink {
  url: string;
  expiresAt: string;
  productName: string;
}