# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StaxVibe AI Graphics marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  customerEmail: string | null;
  amountPaid: number;
  status: string;
  /** @nullable */
  stripeSessionId: string | null;
  /** @nullable */
  downloadToken: string | null;
  createdAt: string;
}