# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StaxVibe AI Graphics marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  featured?: boolean;
  /** @nullable */
  badge?: string | null;
}